Split endpoint validation module
